Move more files out of API folder to respective directories
Move constants to `lib/`.
`api/` directory needs to temporarily stay in `imports/` folder as it's code
that is being used in both `client/` and `server/` files. Meteor does not give
access to server side code in the client directory and is not importable.
Reference: https://github.com/meteor/guide/issues/491#issuecomment-229224828
