+++ client/containers/dashboard.jsx
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
-import Entries from '/imports/api/entries/entries.js';
+import Entries from '/imports/api/entries';
 import Dashboard from '/client/pages/dashboard';
 
 // eslint-disable-next-line no-undef
