+++ client/pages/login.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link, Redirect } from 'react-router-dom';
 import { Alert, Container, Button, Form, Row, Col } from 'react-bootstrap';
 import { Meteor } from 'meteor/meteor';
-import { EMAIL_REGEX } from '/imports/constants/regex';
+import { EMAIL_REGEX } from '/lib/constants/regex';
 import { AuthContext } from '../hoc/AuthProvider';
 class Login extends React.Component {
   constructor(props) {
