+++ imports/api/entries/server/methods.js
@@ -1,3 +1,3 @@
 /* eslint-disable no-unused-vars */
 import { Meteor } from 'meteor/meteor';
-import Entries from '/imports/api/entries/entries.js';
+import Entries from '/imports/api/entries';
