+++ imports/api/entries/server/publications.js
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor';
-import Entries from '/imports/api/entries/entries.js';
+import Entries from '/imports/api/entries';
 
 Meteor.publish('entries.board', function() {
   return Entries.find({}, { limit: 100 });
